refactor(frontend): move components into frontend app src directory
