+++ apps/frontend/docs/architecture-dm.md
