+++ apps/frontend/docs/architecture-eo.md
