+++ apps/frontend/public/project-nova.jpg
