+++ apps/frontend/public/realidea-system.png
