+++ apps/frontend/public/stellar.png
