+++ apps/frontend/src/components/common/FavoriteButton.tsx
