+++ apps/frontend/src/components/common/SharedMessageBar.tsx
