+++ apps/frontend/src/components/directory/RomDirectoryAside.tsx
