+++ apps/frontend/src/components/sharedPageState/SharedMessageBarContainer.tsx
@@ -1,5 +1,5 @@
-import { SharedMessageBar } from "../common/SharedMessageBar";
-import { useSharedPageState } from "../../hooks/useSharedPageState";
+import { SharedMessageBar } from '../common/SharedMessageBar';
+import { useSharedPageState } from '../../hooks/useSharedPageState';
 
 export default function SharedMessageBarContainer() {
   const { state, actions } = useSharedPageState();
