+++ apps/frontend/src/hooks/useFavorites.ts
