+++ apps/frontend/src/hooks/useSearchFilter.ts
