+++ apps/frontend/src/hooks/useSharedPageState.ts
