+++ apps/frontend/src/layouts/Footer.tsx
