+++ apps/frontend/src/layouts/Navbar.tsx
